requirements.txt
apache-airflow-providers-ssh
pandas==1.2.4
numpy==1.20.2
paramiko==2.7.2